+++ Raytracer/Comonents/Basic/Vector.cpp
@@ -1,8 +1,8 @@
 #include "Vector.h"
 
-Vector::Vector() { x = 0; y = 0; z = 0; }
+Vector::Vector() { x = 0.0f; y = 0.0f; z = 0.0f; }
 
-Vector::Vector(double a, double b, double c) {
+Vector::Vector(float a, float b, float c) {
 	x = a;
 	y = b;
 	z = c;
@@ -16,7 +16,7 @@ void Vector::operator = (const Vector& other) {
 
 
 // Vector-Value Operations
-Vector Vector::operator * (const double value) {
+Vector Vector::operator * (const float value) {
 	return Vector(x * value, y * value, z * value);
 }
 
@@ -38,12 +38,12 @@ Vector Vector::operator / (const Vector& other) {
 	return Vector((x / other.x), (y / other.y), (z / other.z));
 }
 
-double Vector::magnitude() {
+float Vector::magnitude() {
 	return sqrt((x * x) + (y * y) + (z * z));
 }
 
 Vector Vector::normalize() {
-	double mag = this->magnitude();
+	float mag = this->magnitude();
 	return Vector(x / mag, y / mag, z / mag);
 }
